backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
Flask-PyMongo==2.3.0
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
tqdm==4.66.1
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2